pyproject.toml
[project]
name = "edubuilder"
version = "0.1.0"
description = "Educational AI Platform API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "streamlit>=1.35.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.4",
    "requests>=2.32.3",
    "sqlmodel>=0.0.18",
    "alembic>=1.13.2",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.4",
    "anyio>=4.4.0",
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "schemathesis>=4.0.0",
    "email-validator>=2.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "anyio: tests that use AnyIO async support",
]

[tool.alembic]
script_location = "alembic"
